feat : 내가 등록한 매칭 상세보기 기능 수정
