+++ src/main/java/com/example/demo/matching/service/MatchingService.java
@@ -1,8 +1,6 @@
 package com.example.demo.matching.service;
 
-import com.example.demo.entity.SiteUser;
 import com.example.demo.matching.dto.MatchingDetailDto;
-import com.example.demo.matching.dto.MatchingMyHostedDto;
 import com.example.demo.matching.dto.MatchingPreviewDto;
 
 import java.util.List;
@@ -13,5 +11,4 @@ public interface MatchingService {
     public void delete(Long matchingId);
     public List<MatchingPreviewDto> getList();
     public MatchingDetailDto getDetail(Long matchingId);
-    List<MatchingMyHostedDto> getBySiteUser(SiteUser siteUser);
 }
