+++ src/main/java/com/example/demo/matching/service/MatchingServiceImpl.java
@@ -1,15 +1,10 @@
 package com.example.demo.matching.service;
 
-import com.example.demo.entity.Matching;
-import com.example.demo.entity.SiteUser;
 import com.example.demo.matching.dto.MatchingDetailDto;
-import com.example.demo.matching.dto.MatchingMyHostedDto;
 import com.example.demo.matching.dto.MatchingPreviewDto;
 import com.example.demo.matching.repository.MatchingRepository;
 import java.util.List;
-import java.util.stream.Collectors;
 
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -42,18 +37,5 @@ public class MatchingServiceImpl implements MatchingService {
     @Override
     public MatchingDetailDto getDetail(Long matchingId) {
         return null;
-    }
-
-    @Override
-    public List<MatchingMyHostedDto> getBySiteUser(SiteUser siteUser) {
-        List<Matching> matchingList = matchingRepository.findBySiteUser(siteUser);
-
-        if (matchingList != null && !matchingList.isEmpty()) {
-            return matchingList.stream()
-                    .map(MatchingMyHostedDto::fromEntity)
-                    .collect(Collectors.toList());
-        } else {
-            throw new EntityNotFoundException("No matching data found for user with ID: " + siteUser.getId());
-        }
     }
 }
