+++ src/main/java/com/example/demo/siteuser/controller/SiteUserInfoController.java
@@ -1,7 +1,9 @@
 package com.example.demo.siteuser.controller;
 
+import com.example.demo.entity.SiteUser;
 import com.example.demo.matching.repository.MatchingRepository;
 import com.example.demo.repository.SiteUserRepository;
+import com.example.demo.siteuser.dto.MatchingMyHostedDto;
 import com.example.demo.siteuser.dto.SiteUserInfoDto;
 import com.example.demo.siteuser.dto.SiteUserMyInfoDto;
 import com.example.demo.siteuser.service.SiteUserInfoService;
@@ -10,6 +12,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 
 @RequiredArgsConstructor
@@ -40,6 +44,17 @@ public class SiteUserInfoController {
 
         if (siteUserMyInfoDto != null) {
             return new ResponseEntity<>(siteUserMyInfoDto, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @GetMapping("/my-page/hosted/{siteUser}")
+    public ResponseEntity<List<MatchingMyHostedDto>> getBySiteUser(@PathVariable(value = "siteUser") SiteUser siteUser) {
+        List<MatchingMyHostedDto> matchingMyHostedDtos = siteUserInfoService.getBySiteUser(siteUser);
+
+        if (!matchingMyHostedDtos.isEmpty()) {
+            return new ResponseEntity<>(matchingMyHostedDtos, HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
