+++ src/main/java/com/example/demo/siteuser/service/SiteUserInfoServiceImpl.java
@@ -1,6 +1,9 @@
 package com.example.demo.siteuser.service;
 
+import com.example.demo.entity.Matching;
 import com.example.demo.entity.SiteUser;
+import com.example.demo.siteuser.dto.MatchingMyHostedDto;
+import com.example.demo.matching.repository.MatchingRepository;
 import com.example.demo.repository.SiteUserRepository;
 import com.example.demo.siteuser.dto.SiteUserInfoDto;
 import com.example.demo.siteuser.dto.SiteUserMyInfoDto;
@@ -8,13 +11,18 @@ import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
 @Service
 public class SiteUserInfoServiceImpl implements SiteUserInfoService {
     private final SiteUserRepository siteUserRepository;
+    private final MatchingRepository matchingRepository;
 
     @Autowired
-    public SiteUserInfoServiceImpl(SiteUserRepository siteUserRepository) {
+    public SiteUserInfoServiceImpl(SiteUserRepository siteUserRepository, MatchingRepository matchingRepository) {
         this.siteUserRepository = siteUserRepository;
+        this.matchingRepository = matchingRepository;
     }
 
     @Override
@@ -29,5 +37,18 @@ public class SiteUserInfoServiceImpl implements SiteUserInfoService {
         SiteUser siteUser = siteUserRepository.findById(userId)
                 .orElseThrow(() -> new EntityNotFoundException("User not found with ID: " + userId));
         return SiteUserMyInfoDto.fromEntity(siteUser);
+    }
+
+    @Override
+    public List<MatchingMyHostedDto> getBySiteUser(SiteUser siteUser) {
+        List<Matching> matchingList = matchingRepository.findBySiteUser(siteUser);
+
+        if (matchingList != null && !matchingList.isEmpty()) {
+            return matchingList.stream()
+                    .map(MatchingMyHostedDto::fromEntity)
+                    .collect(Collectors.toList());
+        } else {
+            throw new EntityNotFoundException("No matching data found for user with ID: " + siteUser.getId());
+        }
     }
 }
